correct some error in tests due to merge
